Create event list package, move relevant Classes there and rename laodetailfragment to a more appropriated EventListFragment
