+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/ui/detail/event/election/fragments/CastVoteFragment.java
@@ -15,9 +15,11 @@ import com.github.dedis.popstellar.model.network.method.message.data.election.Pl
 import com.github.dedis.popstellar.model.objects.Election;
 import com.github.dedis.popstellar.model.objects.view.LaoView;
 import com.github.dedis.popstellar.repository.ElectionRepository;
-import com.github.dedis.popstellar.ui.detail.*;
+import com.github.dedis.popstellar.ui.detail.LaoDetailActivity;
+import com.github.dedis.popstellar.ui.detail.LaoDetailViewModel;
 import com.github.dedis.popstellar.ui.detail.event.election.ZoomOutTransformer;
 import com.github.dedis.popstellar.ui.detail.event.election.adapters.CastVoteViewPagerAdapter;
+import com.github.dedis.popstellar.ui.detail.event.eventlist.EventListFragment;
 import com.github.dedis.popstellar.utility.error.UnknownElectionException;
 import com.github.dedis.popstellar.utility.error.UnknownLaoException;
 
@@ -160,7 +162,7 @@ public class CastVoteFragment extends Fragment {
                     setCurrentFragment(
                         getParentFragmentManager(),
                         R.id.fragment_lao_detail,
-                        LaoDetailFragment::newInstance);
+                        EventListFragment::newInstance);
                     // Toast ? + send back to election screen or details screen ?
                     Toast.makeText(requireContext(), "vote successfully sent !", Toast.LENGTH_LONG)
                         .show();
