+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/ui/detail/event/eventlist/EventListDivider.java
@@ -1,4 +1,4 @@
-package com.github.dedis.popstellar.ui.detail.event;
+package com.github.dedis.popstellar.ui.detail.event.eventlist;
 
 import android.content.Context;
 import android.graphics.*;
