+++ fe2-android/app/src/test/ui/robolectric/com/github/dedis/popstellar/ui/detail/EventListFragmentTest.java
@@ -18,8 +18,6 @@ import com.github.dedis.popstellar.model.network.method.message.data.rollcall.Cr
 import com.github.dedis.popstellar.model.objects.Channel;
 import com.github.dedis.popstellar.model.objects.Lao;
 import com.github.dedis.popstellar.model.objects.security.*;
-import com.github.dedis.popstellar.model.objects.view.LaoView;
-import com.github.dedis.popstellar.model.qrcode.ConnectToLao;
 import com.github.dedis.popstellar.repository.LAORepository;
 import com.github.dedis.popstellar.repository.remote.GlobalNetworkManager;
 import com.github.dedis.popstellar.repository.remote.MessageSender;
@@ -62,7 +60,7 @@ import static org.mockito.Mockito.when;
 @LargeTest
 @HiltAndroidTest
 @RunWith(AndroidJUnit4.class)
-public class LaoDetailFragmentTest {
+public class EventListFragmentTest {
 
   private static final String LAO_NAME = "LAO";
   private static final KeyPair KEY_PAIR = Base64DataUtils.generateKeyPair();
